// ClientApp_old/src/services/openapi/api.ts
export interface AddGoalRequest {
    /**
     * 
     * @type {number}
     * @memberof AddGoalRequest
     */
    'value': number;
    /**
     * 
     * @type {GoalType}
     * @memberof AddGoalRequest
     */
    'type': GoalType;
    /**
     * 
     * @type {GoalPeriod}
     * @memberof AddGoalRequest
     */
    'perPeriod': GoalPeriod;
    /**
     * 
     * @type {string}
     * @memberof AddGoalRequest
     */
    'startDate': string;
    /**
     * 
     * @type {boolean}
     * @memberof AddGoalRequest
     */
    'force'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof AddGoalRequest
     */
    'endDate'?: string | null;
}
// ...
export const GoalPeriod = {
    Day: 'Day',
    Week: 'Week',
    Month: 'Month'
} as const;

export type GoalPeriod = typeof GoalPeriod[keyof typeof GoalPeriod];


/**
 * 
 * @export
 * @enum {string}
 */

export const GoalType = {
    Calories: 'Calories',
    Water: 'Water'
} as const;

export type GoalType = typeof GoalType[keyof typeof GoalType];
// ...
export interface UpdateGoalRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateGoalRequest
     */
    'id': string;
    /**
     * 
     * @type {GoalType}
     * @memberof UpdateGoalRequest
     */
    'type': GoalType;
    /**
     * 
     * @type {number}
     * @memberof UpdateGoalRequest
     */
    'value': number;
    /**
     * 
     * @type {GoalPeriod}
     * @memberof UpdateGoalRequest
     */
    'perPeriod': GoalPeriod;
    /**
     * 
     * @type {string}
     * @memberof UpdateGoalRequest
     */
    'startDate': string;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateGoalRequest
     */
    'force'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateGoalRequest
     */
    'endDate'?: string | null;
}
// ...
export class GoalsApi extends BaseAPI {
    /**
     * 
     * @param {AddGoalRequest} addGoalRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof GoalsApi
     */
    public addGoal(addGoalRequest: AddGoalRequest, options?: RawAxiosRequestConfig) {
        return GoalsApiFp(this.configuration).addGoal(addGoalRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {GoalType} goalType 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof GoalsApi
     */
    public getActiveGoal(goalType: GoalType, options?: RawAxiosRequestConfig) {
        return GoalsApiFp(this.configuration).getActiveGoal(goalType, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateGoalRequest} updateGoalRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof GoalsApi
     */
    public updateGoal(updateGoalRequest: UpdateGoalRequest, options?: RawAxiosRequestConfig) {
        return GoalsApiFp(this.configuration).updateGoal(updateGoalRequest, options).then((request) => request(this.axios, this.basePath));
    }
}